backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic[email]==2.11.7
python-dotenv==1.0.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.28.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1 # forcce version
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
httpx==0.27.2
python-memcached==1.61
pydantic-settings==2.10.1
feedparser==6.0.11
requests==2.32.4
beautifulsoup4==4.13.4
apscheduler==3.10.4
croniter==1.4.1
alembic==1.13.1
aiosqlite==0.20.0
crawl4ai==0.7.2
pytube==15.0.0
# --- Dev dependencies ---
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.27.2
black==23.7.0
isort==5.12.0
mypy==1.5.1
flake8==6.1.0
pytest-cov==6.2.1
pytest-html==4.1.1
boto3==1.39.14
Pillow==10.4.0